Initialised Renderer in Application object as m_Renderer.
Provided m_Renderer with some resizing ability by giving control of glViewport to it.

When the Window resizes and the application object reeives the OnWindowResize event, this information gets passed onto m_Renderer

Currently no accounting has been made for shape geometry warping.

Redesigned Renderer's init to not always require a VAO and ShaderProgram to init but instead a width and height  set of values are required. Checks are made in the init process to see if current VAO/ShaderProgram are nullptrs